feat: sortWorkBlocks 함수 분리
